feat: fix  exisiting event loop problem (#52)
The current implementation creates a new event loop even when one
already exists.
But asyncio does not allow more than one event loop in the same thread,
so this fails in those cases.
For example, FastAPI and Jupyter Notebook already run an event loop, so
they cannot work with pyopenems as-is.

This PR adds a simple mechanism to handle the event loop.

It switches behavior based on whether an event loop is already running:

- No running loop: create a new event loop and run the task there.
- Loop already running: start a new thread and run the task in that
thread.